--- a/config/config.go
+++ b/config/config.go
@@ -240,10 +240,19 @@ type Privacy struct {
 	LMT  LMT
 }
 
+type VendorListFetcher struct {
+	// Max concurrent request for downloading the latest version of the vendor list from a specific major version
+	MaxConcurrencyInitFetchLatestVersion int `mapstructure:"max_concurrency_init_fetch_latest_version"`
+
+	// Max concurrent request for downloading every specific version of the vendor list from its first version to its latest version
+	MaxConcurrencyInitFetchSpecificVersion int `mapstructure:"max_concurrency_init_fetch_specific_version"`
+}
+
 type GDPR struct {
-	Enabled                 bool         `mapstructure:"enabled"`
-	HostVendorID            int          `mapstructure:"host_vendor_id"`
-	DefaultValue            string       `mapstructure:"default_value"`
+	Enabled      bool   `mapstructure:"enabled"`
+	HostVendorID int    `mapstructure:"host_vendor_id"`
+	DefaultValue string `mapstructure:"default_value"`
+	// TODO: migrate the fetcher timeout to VendorListFetcher
 	Timeouts                GDPRTimeouts `mapstructure:"timeouts_ms"`
 	NonStandardPublishers   []string     `mapstructure:"non_standard_publishers,flow"`
 	NonStandardPublisherMap map[string]struct{}
@@ -255,6 +264,9 @@ type GDPR struct {
 	// to DefaultValue
 	EEACountries    []string `mapstructure:"eea_countries"`
 	EEACountriesMap map[string]struct{}
+
+	// VendorListFetcher - configuration for fetching the vendor list from a remote source.
+	VendorListFetcher VendorListFetcher `mapstructure:"vendorlist_fetcher"`
 }
 
 func (cfg *GDPR) validate(v *viper.Viper, errs []error) []error {
@@ -1140,6 +1152,8 @@ func SetupViper(v *viper.Viper, filename string, bidderInfos BidderInfos) {
 		"FIN", "FRA", "GUF", "DEU", "GIB", "GRC", "GLP", "GGY", "HUN", "ISL", "IRL", "IMN", "ITA", "JEY", "LVA",
 		"LIE", "LTU", "LUX", "MLT", "MTQ", "MYT", "NLD", "NOR", "POL", "PRT", "REU", "ROU", "BLM", "MAF", "SPM",
 		"SVK", "SVN", "ESP", "SWE", "GBR"})
+	v.SetDefault("gdpr.vendorlist_fetcher.max_concurrency_init_fetch_latest_version", 1) // by default one download at a time
+	v.SetDefault("gdpr.vendorlist_fetcher.max_concurrency_init_fetch_latest_version", 1) // by default one download at a time
 	v.SetDefault("ccpa.enforce", false)
 	v.SetDefault("lmt.enforce", true)
 	v.SetDefault("currency_converter.fetch_url", "https://cdn.jsdelivr.net/gh/prebid/currency-file@1/latest.json")