*Problem*
The current implementation of downloading vendor list in preload stage is executed sequentially for individual vendor file - it costs 5  to 10 seconds to download all ~300 vendor files on my local machine ( Apple M1 Pro, 32GB memory).

** Solution **
1. the solution is to optimize the preload procedure by applying concurrency which is configurable by the end user. The end user can tune the number depending on their own spec of the machine.

2. in my `pbs.yaml` config , i set the following value which i observed almost **5x** performance improvement ( ~500 to 600 ms ), after 20, the performance increase is diminishing 

```
gdpr:
  default_value: "0" # disable gdpr, explicitly specifying a default value is a requirement in prebid server config
  vendorlist_fetcher:
    max_concurrency_init_fetch_latest_version: 1
    max_concurrency_init_fetch_specific_version: 1
  timeouts_ms:
    init_vendorlist_fetches: 30000
    active_vendorlist_fetch: 30000

``` 


you are welcome to play around with the solution on your own setting and i highly recommend you see the number yourself.