@bsardo 2026-09-01
Thank you for the contribution and for supporting the Prebid Server project. We're closing this PR as the issue will be resolved as part of the active work on #4860.